ds1 editor: moved ds1 widget to its own package
